Ca-DA Refactoring changes.
