+++ src/main/java/uk/gov/hmcts/reform/prl/enums/State.java
@@ -9,8 +9,10 @@ import lombok.RequiredArgsConstructor;
 @Getter
 public enum State {
 
-    @JsonProperty("AWAITING_SUBMISSION_TO_HMCTS")
-    AWAITING_SUBMISSION_TO_HMCTS("AWAITING_SUBMISSION_TO_HMCTS");
+    @JsonProperty("AWAITING_C100_SUBMISSION_TO_HMCTS")
+    AWAITING_C100_SUBMISSION_TO_HMCTS("AWAITING_C100_SUBMISSION_TO_HMCTS"),
+    @JsonProperty("AWAITING_FL401_SUBMISSION_TO_HMCTS")
+    AWAITING_FL401_SUBMISSION_TO_HMCTS("AWAITING_FL401_SUBMISSION_TO_HMCTS");
 
     private final String value;
