+++ src/test/java/uk/gov/hmcts/reform/prl/services/RequestUpdateCallbackServiceTest.java
@@ -25,7 +25,7 @@ import static org.mockito.Mockito.when;
 public class RequestUpdateCallbackServiceTest {
 
     private final String jurisdiction = "PRIVATELAW";
-    private final String caseType = "C100";
+    private final String caseType = "PRLAPPS";
     private final Long caseId = 1234567887654321L;
     private final String eventName = "paymentSuccessCallback";
     private final String userToken = "Bearer testToken";
